Change float to double to fix precision problem

+++ Source/Runtime/Private/RHIOpenGL/RGLInternal.h
@@ -139,7 +139,7 @@ struct GLContext {
     GLuint                  defaultFramebuffer;
     GLuint                  defaultVAO;
     ImGuiContext *          imGuiContext;
-    float                   imGuiLastTime;
+    double                  imGuiLastTime;
 };
 
 struct GLStencilState {
